Converter 구조체 생성
